% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql2.R
\name{cql2}
\alias{cql2}
\alias{cql2_to_text}
\alias{cql2_to_json}
\title{Convert expression to CQL2 format}
\usage{
cql2_to_text(expr)

cql2_to_json(expr)
}
\arguments{
\item{expr}{An R expression to be represented in CQL2}
}
\description{
Convert expression to CQL2 format
}
\examples{

cql2_to_text(datetime >= date("2020-01-01"))

cql2_to_json(`eo:cloud_cover` < 0.2)

}
