% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql2.R
\name{cql2}
\alias{cql2}
\alias{cql2_text}
\alias{cql2_json}
\title{Convert R expressions to CQL2}
\usage{
cql2_text(expr)

cql2_json(expr)
}
\arguments{
\item{expr}{An R expression to be represented in CQL2}
}
\description{
These functions convert R expressions to CQL2 standard (text or JSON).
}
\examples{
# basic cql2 examples
cql2_text("This is a literal string.")
cql2_text("Via dell'Avvento")
cql2_json("Via dell'Avvento")
cql2_text(-100)
cql2_json(-100)
cql2_text(3.14159)
cql2_text(TRUE)
cql2_text(FALSE)
cql2_text(time("1969-07-20T20:17:40Z")) # it also works with timestamp()
cql2_json(time("1969-07-20T20:17:40Z"))
cql2_text(date("1969-07-20"))
cql2_json(date("1969-07-20"))
cql2_text(interval("1969-07-16", "1969-07-24"))
cql2_text(interval("1969-07-16T05:32:00Z", "1969-07-24T16:50:35Z"))
cql2_text(interval("2019-09-09", ".."))
cql2_json(interval("1969-07-16", "1969-07-24"))
cql2_json(interval("1969-07-16T05:32:00Z", "1969-07-24T16:50:35Z"))
cql2_json(interval("2019-09-09", ".."))
cql2_text(city == "Toronto")
cql2_json(city == "Toronto")
cql2_text(balance - 150.0 > 0)
cql2_json(balance - 150.0 > 0)
cql2_text(updated >= date("1970-01-01"))
cql2_json(updated >= date("1970-01-01"))
cql2_text(!is_null(geometry))
cql2_json(!is_null(geometry))
poly_sf <- sfheaders::sf_polygon(matrix(c(0,0,0,0,1,1), ncol = 2))
cql2_text(s_intersects({{poly_sf}}, geometry))
cql2_json(s_intersects({{poly_sf}}, geometry))
cql2_text(s_crosses(geometry, {{poly_sf}}))
cql2_json(s_crosses(geometry, {{poly_sf}}))
}
